Add SAR frame properties
SAR in NTSC/PAL is a mess.  There are several relevant standards.  Sampling is
specified in BT.601 and SMPTE 244M/EBU 3280, but none define active line
width and blanking.  However, SMPTE RP 187 (plus Annex A.4) and EBU R92 do
address production apertures and aspect ratios.

525-line NTSC
NTSC active line width is debatable.  SMPTE 170M (63+5/9 - 1.5 - 9.2
(+0.3 or -0.2)) and BT.470 (63+5/9 - 10.9 or 52+59/90 ± 0.2) are just within
each other's tolerance.  Poynton uses 52+8/9 which is close to 170M but outside
of BT.470.  SMPTE RP 187 is closer to 170M and Poynton than to BT.470.

Note that while SMPTE RP 187 has been withdrawn, the NTSC section seems
reasonable and ColorBars() follows it.  AVID apparently followed it too, so
that’s a good enough endorsement for me.  RP 187’s PAR is within 0.4% of 170M
anyway, so most of this is academic plus it's only metadata.  You can override
or ignore it.

625-line PAL
PAL active line width is generally accepted to be 52 us.  EBU R92 refers to
BT.470 (52 us).  SMPTE RP 187 defines a unique clean aperture not found
elsewhere, so ColorBars() ignores it and follows R92.